update the evaluations code to have a condition_group evaluation
