+++ src/sentry/workflow_engine/processors/__init__.py
@@ -1,0 +1,6 @@
+__all__ = [
+    "DataConditionEvaluation",
+    "DataConditionGroupEvaluation",
+]
+
+from .evaluations import DataConditionEvaluation, DataConditionGroupEvaluation
