+++ src/sentry/workflow_engine/processors/evaluations/__init__.py
@@ -1,8 +1,10 @@
 __all__ = [
     "DataConditionEvaluation",
     "DataConditionEvaluationException",
+    "DataConditionGroupEvaluation",
     "TriggerResult",
 ]
 
 from .condition import DataConditionEvaluation, DataConditionEvaluationException
+from .condition_group import DataConditionGroupEvaluation
 from .trigger_result import TriggerResult
