+++ src/sentry/workflow_engine/processors/evaluations/base.py
@@ -1,4 +1,7 @@
 from dataclasses import dataclass
+from functools import cached_property
+
+from sentry.workflow_engine.processors.evaluations.trigger_result import TriggerResult
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -9,3 +12,10 @@ class BaseWorkflowEngineEvaluation[R, E]:
 
     result: R | None = None
     error: E | None = None
+
+    @cached_property
+    def outcome(self) -> TriggerResult:
+        return TriggerResult(
+            triggered=(self.result is not None),
+            error=self.error,
+        )
