+++ src/sentry/workflow_engine/processors/evaluations/condition.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from functools import cached_property
 from typing import TYPE_CHECKING, Any
 
 from sentry.workflow_engine.types import ConditionError, DataConditionResult
 
 from .base import BaseWorkflowEngineEvaluation
-from .trigger_result import TriggerResult
 
 if TYPE_CHECKING:
     from sentry.workflow_engine.models.data_condition import DataCondition
@@ -33,17 +31,3 @@ class DataConditionEvaluation(BaseWorkflowEngineEvaluation[DataConditionResult, 
 
     value: Any
     condition: DataCondition
-
-    @cached_property
-    def outcome(self) -> TriggerResult:
-        """
-        TODO - @saponifi3d - The TriggerResult and the BaseWorkflowEngineEvaluation
-        can likely serve the same purpose, looking at the result / errors and providing
-        helpful interactions.
-
-        For now, using the `TriggerResult` to move a little faster through the refactoring.
-        """
-        return TriggerResult(
-            triggered=(self.result is not None),
-            error=self.error,
-        )
